refactor: migrate source files from .mjs to .js with ESM package type
Add "type": "module" to package.json and rename all source files from .mjs to .js so the project is consistently ES module-native. Update rollup.config.js to ESM syntax; dist outputs change from .js to .cjs to make the CJS bundle extension explicit. Add -m flag to qjs invocations in Makefile and report/src/main.js so QuickJS treats the renamed files as modules.
