+++ docker/files/tools/lib/log-parser.js
