+++ docker/files/tools/lib/test-shim.js
