+++ report/action.yml
@@ -13,4 +13,4 @@ inputs:
 
 runs:
   using: node24
-  main: dist/main.js
+  main: dist/main.cjs
