+++ report/src/lib/build-example.js
