+++ setup/action.yml
@@ -25,5 +25,5 @@ inputs:
 
 runs:
   using: node24
-  main: dist/main.js
-  post: dist/post.js
+  main: dist/main.cjs
+  post: dist/post.cjs
